requirements.txt
# requirements.txt
# Core dependencies
r2pipe==1.8.0
rich==13.7.0
fpdf==1.7.2
openai==1.12.0

# API and web
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
stripe==7.12.0

# Database and caching
sqlite3
redis==5.0.1
python-dotenv==1.0.1
cryptography==42.0.2
pyjwt==2.8.0

# Analytics and visualization
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2

# Testing and development
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Optional dependencies for advanced features
angr==9.2.87  # For symbolic execution
capstone==5.0.1  # For disassembly
keystone-engine==0.9.2  # For assembly
unicorn==2.0.1.post1  # For emulation
lief==0.14.1  # For binary manipulation

# Cloud and integration
boto3==1.34.37  # For AWS integration
requests==2.31.0
aiohttp==3.9.3
websockets==12.0

# Security
pycryptodome==3.20.0
argon2-cffi==23.1.0